% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_snRNAseq.R
\name{analyse_snRNAseq}
\alias{analyse_snRNAseq}
\title{analyse snRNAseq output from Parse Biosciences split-pipe}
\usage{
analyse_snRNAseq(
  parse_dge_dir,
  do_filtering = T,
  remove_doublets = T,
  min_cells_per_gene = 5,
  max_nCount_RNA = NULL,
  min_nFeature_RNA = NULL,
  max_nFeature_RNA = NULL,
  max_percent_mito = NULL,
  n_dims = NULL,
  clustering_resolutions = seq(0.1, 0.8, by = 0.1),
  final_clustering_resolution = NULL,
  out_dir = NULL,
  sample_subset = NULL,
  do_timestamp = F,
  do_integration = F,
  integration_col = "sample"
)
}
\arguments{
\item{parse_dge_dir}{Path to Parse Biosciences split-pipe output. This directory must contain \code{DGE.mtx}, \code{all_genes.csv}, and \code{cell_metadata.csv} files.}

\item{do_filtering}{If \code{TRUE}, apply quality control filters to genes and nuclei.}

\item{remove_doublets}{If \code{TRUE}, removes suspected doublet nuclei, as detected by the scDblFinder package.}

\item{min_cells_per_gene}{\emph{Gene filter}. The minimum number of nuclei in which a gene must be present. Genes that are present in very few nuclei are uninformative and unlikely to help in differentiating groups of cells. In general, most genes removed by this filtering will be those not detected in any nucleus.}

\item{max_nCount_RNA}{\emph{Nucleus filter}. The maximum number of transcripts detected per nucleus. An unusually high number of RNA molecules suggests that the nucleus is a multiplet.}

\item{min_nFeature_RNA}{\emph{Nucleus filter}. The minimum number of genes detected per nucleus. An unusually low number of genes suggests that the nucleus has a damaged membrane and is therefore low quality.}

\item{max_nFeature_RNA}{\emph{Nucleus filter}. The maximum number of genes detected per nucleus. An unusually high number of genes suggests that the nucleus is a multiplet.}

\item{max_percent_mito}{\emph{Nucleus filter}. The maximum percentage of mitochondrial transcripts per nucleus. Overrepresentation of mitochondrial transcripts suggests cell death, loss of cytoplasmic RNA, or heightened apoptosis.}

\item{n_dims}{Optional. The dimensionality of the dataset to use for downstream analysis. This is the number of principal components believed to capture the majority of true biological signal in the dataset. This can be decided by consulting the elbow plot. If no value given, dimensionality is calculated using the \code{intrinsicDimensions} package.}

\item{final_clustering_resolution}{Optional. The chosen clustering resolution of the dataset to use for downstream analysis. This is the resolution at which clusters appear to capture true biological groupings of interest in the dataset. This can be decided by consulting the clustering tree. If no value given, downstream analysis will not proceed.}

\item{out_dir}{Optional. Output directory. If no value given, the output will be saved to a path that mirrors the Parse Biosciences split-pipe directory structure (analysis/*).}

\item{sample_subset}{Optional. Vector of sample IDs to subset to.}

\item{do_timestamp}{If \code{TRUE}, will save the output in a time-stamped subdirectory.}

\item{do_integration}{If \code{TRUE}, will integrate the dataset.}

\item{integration_col}{If \code{do_integration} is \code{TRUE}, Seurat metadata column upon which to integrate the dataset.}

\item{cluster_resolutions}{Optional. A vector of clustering resolutions to test. A higher resolution will result in a larger number of communities. Default is 0.1-0.8.}
}
\value{
A Seurat object.
}
\description{
analyse snRNAseq output from Parse Biosciences split-pipe
}
