% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_parse.R
\name{analyse_parse}
\alias{analyse_parse}
\title{Add together two numbers}
\usage{
analyse_parse(
  experiment,
  parse_pipeline_dir = "/camp/project/tracerX/working/VHL_GERMLINE/tidda/parse_pipeline/",
  genome = "hg38",
  parse_analysis_subdir = "/all-well/DGE_filtered/",
  out_dir = NULL,
  sample_subset = NULL,
  min_genes = 3,
  min_cells = 100,
  do_timestamp = F
)
}
\arguments{
\item{experiment}{The name of the experiment directory in the Parse pipeline.}

\item{parse_pipeline_dir}{The full path to the directory where the Parse split-pipe pipeline was run. This should contain genomes/, expdata/, and analysis/ directories.}

\item{genome}{The name of the genome build used as a reference. This will also be the immediate subdirectory of the experiment's analysis.}

\item{parse_analysis_subdir}{The Parse output directory to be used.}

\item{out_dir}{The output directory. If none given, output will save to out/\code{experiment}/\code{genome}/.}

\item{sample_subset}{Sample IDs to subset to.}
}
\value{
A Seurat object
}
\description{
Add together two numbers
}
